+++ lib/guard/rubocop/runner.rb
@@ -92,13 +92,15 @@ module Guard
         text = pluralize(summary[:inspected_file_count], 'file')
         text << ' inspected, '
 
-        text << pluralize(summary[:offense_count], 'offense', no_for_zero: true)
+        offense_count = summary[:offense_count] || summary[:offence_count]
+        text << pluralize(offense_count, 'offense', no_for_zero: true)
         text << ' detected'
       end
 
       def failed_paths
         failed_files = result[:files].reject do |file|
-          file[:offenses].empty?
+          offenses = file[:offenses] || file[:offences]
+          offenses.empty?
         end
         failed_files.map do |file|
           file[:path]